feat: refactor table data hook to accept external pagination and filter parameters
- Removed internal state management (page, limit, searchQuery, statusFilter) from useTableData hook
- Modified useTableData to accept pagination and filter parameters as function arguments
- Moved filter state management (searchQuery, statusFilter) to parent component
- Integrated useTableParams hook for pagination state management before query execution
- Updated useTableState to consume params object instead of individual
